climc: k8s import cluster support
