+++ cmd/climc/shell/k8s/clusters.go
@@ -16,9 +16,6 @@ package k8s
 
 import (
 	"fmt"
-	"io/ioutil"
-	"os"
-	"os/exec"
 
 	"yunion.io/x/jsonutils"
 
@@ -27,63 +24,21 @@ import (
 	o "yunion.io/x/onecloud/pkg/mcclient/options/k8s"
 )
 
-func initCluster() {
-	cmdN := func(suffix string) string {
-		return kubeResourceCmdN("cluster", suffix)
+func initKubeCluster() {
+	cmdN := func(action string) string {
+		return fmt.Sprintf("kubecluster-%s", action)
 	}
-
-	R(&o.ClusterListOptions{}, cmdN("list"), "List k8s infra clusters", func(s *mcclient.ClientSession, args *o.ClusterListOptions) error {
-		result, err := k8s.Clusters.List(s, args.Params())
+	R(&o.ClusterListOptions{}, cmdN("list"), "List k8s clusters", func(s *mcclient.ClientSession, args *o.ClusterListOptions) error {
+		result, err := k8s.KubeClusters.List(s, args.Params())
 		if err != nil {
 			return err
 		}
-		printList(result, k8s.Clusters.GetColumns(s))
-		return nil
-	})
-
-	R(&o.ClusterCreateOptions{}, cmdN("create"), "Create k8s cluster", func(s *mcclient.ClientSession, args *o.ClusterCreateOptions) error {
-		cluster, err := k8s.Clusters.Create(s, args.Params())
-		if err != nil {
-			return err
-		}
-		printObject(cluster)
-		return nil
-	})
-
-	R(&o.ClusterImportOptions{}, cmdN("import"), "Import exists YKE deployed kubernetes cluster", func(s *mcclient.ClientSession, args *o.ClusterImportOptions) error {
-		params, err := args.Params()
-		if err != nil {
-			return err
-		}
-
-		cluster, err := k8s.Clusters.PerformAction(s, args.NAME, "import", params)
-		if err != nil {
-			return err
-		}
-		printObject(cluster)
-		return nil
-	})
-
-	R(&o.ClusterUpdateOptions{}, cmdN("update"), "Update k8s cluster", func(s *mcclient.ClientSession, args *o.ClusterUpdateOptions) error {
-		cluster, err := k8s.Clusters.Update(s, args.NAME, args.Params())
-		if err != nil {
-			return err
-		}
-		printObject(cluster)
-		return nil
-	})
-
-	R(&o.ClusterDeployOptions{}, cmdN("deploy"), "Start deploy a cluster", func(s *mcclient.ClientSession, args *o.ClusterDeployOptions) error {
-		ret, err := k8s.Clusters.PerformAction(s, args.ID, "deploy", args.Params())
-		if err != nil {
-			return err
-		}
-		printObject(ret)
+		printList(result, k8s.KubeClusters.GetColumns(s))
 		return nil
 	})
 
 	R(&o.IdentOptions{}, cmdN("show"), "Show details of a cluster", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
-		result, err := k8s.Clusters.Get(s, args.ID, nil)
+		result, err := k8s.KubeClusters.Get(s, args.ID, nil)
 		if err != nil {
 			return err
 		}
@@ -91,14 +46,75 @@ func initCluster() {
 		return nil
 	})
 
-	R(&o.ClusterDeleteOptions{}, cmdN("delete"), "Delete cluster", func(s *mcclient.ClientSession, args *o.ClusterDeleteOptions) error {
-		ret := k8s.Clusters.BatchDeleteWithParam(s, args.ID, nil, nil)
-		printBatchResults(ret, k8s.Clusters.GetColumns(s))
+	R(&o.KubeClusterCreateOptions{}, cmdN("create"), "Create k8s cluster", func(s *mcclient.ClientSession, args *o.KubeClusterCreateOptions) error {
+		params, err := args.Params()
+		if err != nil {
+			return err
+		}
+		cluster, err := k8s.KubeClusters.Create(s, params)
+		if err != nil {
+			return err
+		}
+		printObject(cluster)
 		return nil
 	})
 
-	R(&o.ClusterKubeconfigOptions{}, cmdN("kubeconfig"), "Generate kubeconfig of a cluster", func(s *mcclient.ClientSession, args *o.ClusterKubeconfigOptions) error {
-		ret, err := k8s.Clusters.PerformAction(s, args.ID, "generate-kubeconfig", args.Params())
+	R(&o.KubeClusterImportOptions{}, cmdN("import"), "Import k8s cluster", func(s *mcclient.ClientSession, args *o.KubeClusterImportOptions) error {
+		params, err := args.Params()
+		if err != nil {
+			return err
+		}
+		cluster, err := k8s.KubeClusters.Create(s, params)
+		if err != nil {
+			return err
+		}
+		printObject(cluster)
+		return nil
+	})
+
+	R(&o.ClusterDeleteOptions{}, cmdN("delete"), "Delete cluster", func(s *mcclient.ClientSession, args *o.ClusterDeleteOptions) error {
+		ret := k8s.KubeClusters.BatchDeleteWithParam(s, args.ID, nil, nil)
+		printBatchResults(ret, k8s.KubeClusters.GetColumns(s))
+		return nil
+	})
+
+	R(&o.KubeClusterAddMachinesOptions{}, cmdN("add-machines"), "Add machines to cluster", func(s *mcclient.ClientSession, args *o.KubeClusterAddMachinesOptions) error {
+		params, err := args.Params()
+		if err != nil {
+			return err
+		}
+		ret, err := k8s.KubeClusters.PerformAction(s, args.ID, "add-machines", params)
+		if err != nil {
+			return err
+		}
+		printObject(ret)
+		return nil
+	})
+
+	R(&o.KubeClusterDeleteMachinesOptions{}, cmdN("delete-machines"), "Delete machines in cluster", func(s *mcclient.ClientSession, args *o.KubeClusterDeleteMachinesOptions) error {
+		params, err := args.Params()
+		if err != nil {
+			return err
+		}
+		ret, err := k8s.KubeClusters.PerformAction(s, args.ID, "delete-machines", params)
+		if err != nil {
+			return err
+		}
+		printObject(ret)
+		return nil
+	})
+
+	R(&o.IdentOptions{}, cmdN("terminate"), "Terminate cluster", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
+		ret, err := k8s.KubeClusters.PerformAction(s, args.ID, "terminate", nil)
+		if err != nil {
+			return err
+		}
+		printObject(ret)
+		return nil
+	})
+
+	R(&o.IdentOptions{}, cmdN("kubeconfig"), "Generate kubeconfig of a cluster", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
+		ret, err := k8s.KubeClusters.GetSpecific(s, args.ID, "kubeconfig", nil)
 		if err != nil {
 			return err
 		}
@@ -110,12 +126,12 @@ func initCluster() {
 		return nil
 	})
 
-	R(&o.IdentOptions{}, cmdN("engineconfig"), "Get kubernetes engine config of a cluster", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
-		ret, err := k8s.Clusters.GetSpecific(s, args.ID, "engine-config", nil)
+	R(&o.IdentOptions{}, cmdN("addons"), "Get addon manifest of a cluster", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
+		ret, err := k8s.KubeClusters.GetSpecific(s, args.ID, "addons", nil)
 		if err != nil {
 			return err
 		}
-		conf, err := ret.GetString("config")
+		conf, err := ret.GetString("addons")
 		if err != nil {
 			return err
 		}
@@ -123,94 +139,39 @@ func initCluster() {
 		return nil
 	})
 
-	R(&o.IdentOptions{}, cmdN("webhookauthurl"), "Get cluster kubernetes api server webhook auth url", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
-		ret, err := k8s.Clusters.GetSpecific(s, args.ID, "webhook-auth-url", nil)
-		if err != nil {
-			return err
-		}
-		url, err := ret.GetString("url")
-		if err != nil {
-			return err
-		}
-		fmt.Println(url)
-		return nil
-	})
-
-	R(&o.ClusterAddNodesOptions{}, cmdN("addnodes"), "Add nodes to cluster", func(s *mcclient.ClientSession, args *o.ClusterAddNodesOptions) error {
-		params, err := args.Params()
-		if err != nil {
-			return err
-		}
-		ret, err := k8s.Clusters.PerformAction(s, args.ID, "add-nodes", params)
-		if err != nil {
-			return err
-		}
-		printObject(ret)
-		return nil
-	})
-
-	R(&o.ClusterDeleteNodesOptions{}, cmdN("delete-nodes"), "Delete nodes in cluster", func(s *mcclient.ClientSession, args *o.ClusterDeleteNodesOptions) error {
-		params, err := args.Params()
-		if err != nil {
-			return err
-		}
-		ret, err := k8s.Clusters.PerformAction(s, args.ID, "delete-nodes", params)
-		if err != nil {
-			return err
-		}
-		printObject(ret)
-		return nil
-	})
-
-	R(&o.IdentOptions{}, cmdN("engineconfig-edit"), "Edit and update kubernetes engine config of a cluster", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
-		ret, err := k8s.Clusters.GetSpecific(s, args.ID, "engine-config", nil)
-		if err != nil {
-			return err
-		}
-		conf, err := ret.GetString("config")
-		if err != nil {
-			return err
-		}
-		tempfile, err := ioutil.TempFile("", fmt.Sprintf("%s-engineconfig", args.ID))
-		if err != nil {
-			return err
-		}
-		defer os.Remove(tempfile.Name())
-
-		if _, err := tempfile.Write([]byte(conf)); err != nil {
-			return err
-		}
-		if err := tempfile.Close(); err != nil {
-			return err
-		}
-
-		cmd := exec.Command("vim", tempfile.Name())
-		cmd.Stdin = os.Stdin
-		cmd.Stdout = os.Stdout
-		if err := cmd.Run(); err != nil {
-			return err
-		}
-
+	R(&o.ClusterK8sVersions{}, cmdN("k8s-versions"), "Get kubernetes deployable versions", func(s *mcclient.ClientSession, args *o.ClusterK8sVersions) error {
 		params := jsonutils.NewDict()
-		configBytes, err := ioutil.ReadFile(tempfile.Name())
+		params.Add(jsonutils.NewString(args.PROVIDER), "provider")
+		ret, err := k8s.KubeClusters.Get(s, "k8s-versions", params)
 		if err != nil {
 			return err
 		}
-		params.Add(jsonutils.NewString(string(configBytes)), "config")
-		result, err := k8s.Clusters.PerformAction(s, args.ID, "update-engine-config", params)
-		if err != nil {
-			return err
-		}
-		conf, err = result.GetString("config")
-		if err != nil {
-			return err
-		}
-		fmt.Println(conf)
+		fmt.Println(ret.String())
 		return nil
 	})
 
-	R(&o.IdentOptions{}, cmdN("sync-config"), "Sync kubernetes cluster by engineconfig", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
-		ret, err := k8s.Clusters.PerformAction(s, args.ID, "sync-config", nil)
+	R(&o.ClusterK8sVersions{}, cmdN("usable-instances"), "Get deploy usable instance", func(s *mcclient.ClientSession, args *o.ClusterK8sVersions) error {
+		params := jsonutils.NewDict()
+		params.Add(jsonutils.NewString(args.PROVIDER), "provider")
+		ret, err := k8s.KubeClusters.Get(s, "usable-instances", params)
+		if err != nil {
+			return err
+		}
+		fmt.Println(ret.String())
+		return nil
+	})
+
+	R(&o.ClusterCheckOptions{}, cmdN("check-system-ready"), "Check system cluster status", func(s *mcclient.ClientSession, args *o.ClusterCheckOptions) error {
+		ret, err := k8s.KubeClusters.PerformClassAction(s, "check-system-ready", nil)
+		if err != nil {
+			return err
+		}
+		fmt.Println(ret.String())
+		return nil
+	})
+
+	R(&o.IdentOptions{}, cmdN("apply-addons"), "Apply base requirements addons", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
+		ret, err := k8s.KubeClusters.PerformAction(s, args.ID, "apply-addons", nil)
 		if err != nil {
 			return err
 		}
@@ -218,12 +179,26 @@ func initCluster() {
 		return nil
 	})
 
-	R(&o.ClusterRestartAgentsOptions{}, cmdN("restart-agent"), "Restart node agents in cluster", func(s *mcclient.ClientSession, args *o.ClusterRestartAgentsOptions) error {
-		params, err := args.Params()
+	R(&o.IdentOptions{}, cmdN("syncstatus"), "Sync cluster status", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
+		ret, err := k8s.KubeClusters.PerformAction(s, args.ID, "syncstatus", nil)
 		if err != nil {
 			return err
 		}
-		ret, err := k8s.Clusters.PerformAction(s, args.ID, "restart-agent", params)
+		printObject(ret)
+		return nil
+	})
+
+	R(&o.IdentOptions{}, cmdN("public"), "Make cluster public", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
+		ret, err := k8s.KubeClusters.PerformAction(s, args.ID, "public", nil)
+		if err != nil {
+			return err
+		}
+		printObject(ret)
+		return nil
+	})
+
+	R(&o.IdentOptions{}, cmdN("private"), "Make cluster private", func(s *mcclient.ClientSession, args *o.IdentOptions) error {
+		ret, err := k8s.KubeClusters.PerformAction(s, args.ID, "private", nil)
 		if err != nil {
 			return err
 		}
