+++ pkg/mcclient/modules/k8s/cluster.go
@@ -19,17 +19,12 @@ import (
 )
 
 var (
-	Clusters     *ResourceManager
 	KubeClusters *ResourceManager
 )
 
 func init() {
-	Clusters = NewResourceManager("kube_cluster", "kube_clusters",
-		NewResourceCols("mode", "k8s_version", "status", "api_endpoint"),
-		NewColumns("is_public"))
 	KubeClusters = NewResourceManager("kubecluster", "kubeclusters",
 		NewResourceCols("cluster_type", "resource_type", "cloud_type", "version", "status", "mode", "provider", "machines"),
 		NewColumns())
-	modules.Register(Clusters)
 	modules.Register(KubeClusters)
 }
